Abstract the NPC creation process.
